Fixes NoClassDefFoundError and NPE in Admob cordova plugin
Apart from one single usage, the plugin currently does not seem to use
any Java 8 language APIs. The usage in question is in Admob.java:

    waitingForReadyCallbackContextResults.forEach()

It compiles successfully with JavaVersion.VERSION_1_8 enabled, however,
it throws a runtime exception on Android SDK < 24 (< v7.0) devices.

    java.lang.NoClassDefFoundError: admob.plugin.-$$Lambda$AdMob$GO15ms8Jdvu2s5w9ooldE6Mb0qs

It does not seem that the methdo is supported at least in the Android documentation at
https://developer.android.com/studio/write/java8-support.html. This commit replaces the
forEach() call with a simple for-loop iteration.

The second part of the commit replaces nullification of the array with a clear() method call.

    waitingForReadyCallbackContextResults = null;

It's not quite clear why this was set to null in the first place, however, it causes a NPE
in a case where the cordova 'ready' event gets triggered twice. The event could be triggered
several times for the case where the WebView loads html files via location.href = '...'